Address PR comments
Remove dbus-python from the pipfile, no longer needed. At the same time
update the dependencies away from locked versions. Was having some problems
with impossible installs for dependencies.

Remove unneeded #! lines and update copyrights

Added punctuation and clarified YANG revisions and descriptions

Removed some unneeded + characters inside parenthesis

Removed sanity marks as all unit tests are run as part of the build

Removed tests for constants

Fixed some white space around vif_sanitizing and test_garp

Removed unneeded classes and functions in mocks

Fixed spelling mistakes in dbus path mocks and simplified if statement for
checking the path or name is what the mock expects

Simplified writing of some mock files saving variables

Removed monkeypatch and used setattr for mocks instead

Removed unit_state method from ProcessControl, nothing used it

Simplified is_running and get_agent_x_socket methods in Process control

Simplified determining address family type in VrrpConnection

Changed return type of garp in VrrpConnection to None and stop return of
dictionary

Updated return type of send_garp in vyatta_vrrp_vci.py to None as it wasn't
returning anything

Simplified rfc_intf_map() function saving a variable

Updated some comments with more punctuation and better descriptions

Instead of using a series of or toggles for show commands used in list

Change from float to decimal type, is more modern

+++ vyatta/vrrp_vci/keepalived/vrrp.py
@@ -1,11 +1,10 @@
-#! /usr/bin/env python3
-
 # Copyright (c) 2020 AT&T Intellectual Property.
 # All rights reserved.
 # SPDX-License-Identifier: GPL-2.0-only
 
 
 import logging
+from decimal import Decimal
 from typing import Dict
 
 import vyatta.vrrp_vci.keepalived.util as util
@@ -52,7 +51,7 @@ class VrrpGroup:
                 group_config[util.YANG_V2_ADVERT_INT]
             del self._group_config[util.YANG_V2_ADVERT_INT]
         elif util.YANG_V3_ADVERT_INT in self._group_config:
-            fast_advertise: float = \
+            fast_advertise: Decimal = \
                 group_config[util.YANG_V3_ADVERT_INT] / 1000
             if group_config[util.YANG_V3_ADVERT_INT] % 1000 == 0:
                 self._group_config[util.CONFIG_ADVERT] = int(fast_advertise)
